Implement basic password manager terminal app
Starts by asking for a password that is used with a hardcoded IV to
provide AES CBC (code block chaining) encryption and decryption over a
file (~/.__passpass) that keeps all passwords.

A minimal repl is supplied to list, add and delete entries, and to
quit the program.

When adding an entry, auto-generated passwords are suggested until the
target site accepts them.